+++ groups_manager/tests/test_models.py
@@ -2,6 +2,7 @@ from copy import deepcopy
 import re
 import sys
 
+from django.contrib.auth import get_user_model
 from django.contrib.auth.models import Group as DjangoGroup
 from django.test import TestCase
 
@@ -10,7 +11,7 @@ from groups_manager import models, exceptions_gm
 
 GROUPS_MANAGER_MOCK = {
     'AUTH_MODELS_SYNC': True,
-    'AUTH_MODELS_GET_OR_CREATE': False,
+    'AUTH_MODELS_GET_OR_CREATE': True,
     'GROUP_NAME_PREFIX': 'DGM_',
     'GROUP_NAME_SUFFIX': '_$$random',
     'USER_USERNAME_PREFIX': 'DGM_',
@@ -69,6 +70,16 @@ class TestMember(TestCase):
         member = models.Member.objects.create(first_name='Caio', last_name='Mario')
         self.assertEqual(member.username, member.django_user.username)
 
+    def test_member_save_to_existing_user(self):
+        from groups_manager import settings
+        empty_ps = deepcopy(GROUPS_MANAGER_MOCK)
+        empty_ps['USER_USERNAME_PREFIX'] = ''
+        empty_ps['USER_USERNAME_SUFFIX'] = ''
+        settings.GROUPS_MANAGER = empty_ps
+        django_user = get_user_model().objects.create(username='caio_mario')
+        member = models.Member.objects.create(first_name='Caio', last_name='Mario')
+        self.assertEqual(django_user, member.django_user)
+
     def test_has_perm_exception(self):
         member = models.Member(first_name='Caio', last_name='Mario')
         with self.assertRaises(exceptions_gm.MemberDjangoUserSyncError):
@@ -103,6 +114,13 @@ class TestGroupType(TestCase):
         self.group_type.save()
         self.assertEqual(self.group_type.codename, 'organization')
 
+    def test_group_type_groups_reverse(self):
+        self.group_type.save()
+        g1 = models.Group.objects.create(name='Group 1', group_type=self.group_type)
+        self.assertEqual(list(self.group_type.groups_manager_group_set.all()), [g1])
+        # Deprecated
+        self.assertEqual(list(self.group_type.groups.all()), [g1])
+
 
 class TestGroupEntity(TestCase):
 
@@ -125,6 +143,14 @@ class TestGroupEntity(TestCase):
         self.group_entity.codename = 'organization-partner'
         self.group_entity.save()
         self.assertEqual(self.group_entity.codename, 'organization-partner')
+
+    def test_group_entity_groups_reverse(self):
+        self.group_entity.save()
+        g1 = models.Group.objects.create(name='Group 1')
+        g1.group_entities.add(self.group_entity)
+        self.assertEqual(list(self.group_entity.groups_manager_group_set.all()), [g1])
+        # Deprecated
+        self.assertEqual(list(self.group_entity.groups.all()), [g1])
 
 
 class TestGroup(TestCase):
@@ -206,6 +232,16 @@ class TestGroup(TestCase):
         subgroup.save()
         self.assertEqual(subgroup.django_group.name, 'Main Group-Sub group')
 
+    def test_group_save_to_existing_group(self):
+        from groups_manager import settings
+        empty_ps = deepcopy(GROUPS_MANAGER_MOCK)
+        empty_ps['GROUP_NAME_PREFIX'] = ''
+        empty_ps['GROUP_NAME_SUFFIX'] = ''
+        settings.GROUPS_MANAGER = empty_ps
+        django_group = models.DjangoGroup.objects.create(name='Main Group')
+        group = models.Group.objects.create(name='Main Group')
+        self.assertEqual(django_group, group.django_group)
+
     def test_group_save_django_group_change_name(self):
         from groups_manager import settings
         settings.GROUPS_MANAGER = deepcopy(GROUPS_MANAGER_MOCK)
@@ -215,6 +251,18 @@ class TestGroup(TestCase):
         group.name = 'Test change'
         group.save()
         self.assertEqual(group.django_group.name, group.name)
+
+    def test_group_save_django_group_name_length(self):
+        """Test group name with more than 80 chars."""
+        from groups_manager import settings
+        settings.GROUPS_MANAGER = deepcopy(GROUPS_MANAGER_MOCK)
+        settings.GROUPS_MANAGER['GROUP_NAME_PREFIX'] = 'ten_chars_'
+        settings.GROUPS_MANAGER['GROUP_NAME_SUFFIX'] = '_ten_chars'
+        parent = models.Group.objects.create(name='Main_Group_2')  # 12
+        group = models.Group.objects.create(name='a_b_c_d_e' * 6, parent=parent)  # 10
+        group.save()
+        long_name_without_parent = 'ten_chars_oup_2-%s_ten_chars' % ('a_b_c_d_e' * 6)
+        self.assertEqual(group.django_group.name, long_name_without_parent)
 
     def test_nested_users(self):
         from groups_manager import settings
@@ -349,3 +397,13 @@ class TestGroupMember(TestCase):
         name = main.django_group.name
         main.delete()
         self.assertEqual(len(DjangoGroup.objects.filter(name=name)), 0)
+
+    def test_member_groups_reverse(self):
+        m1 = models.Member.objects.create(first_name='Caio', last_name='Mario')
+        g1 = models.Group.objects.create(name='Group 1')
+        g2 = models.Group.objects.create(name='Group 2')
+        models.GroupMember.objects.create(group=g1, member=m1)
+        models.GroupMember.objects.create(group=g2, member=m1)
+        self.assertEqual(list(m1.groups_manager_group_set.all()), [g1, g2])
+        # Deprecated
+        self.assertEqual(list(m1.groups.all()), [g1, g2])